Adjusted the imports according to the renamings. Moved the mask utils to services (they were in tools)
