+++ docs/TODO.md
@@ -1,8 +1,32 @@
 # TODO — HTAS cleanup
 
 Clean up to keep the architecture consistent: avoid repetition and dead code,
-and keep a solid division of roles across the layers (client → dispatch →
-handlers → accessors).
+and keep a solid division of roles across the 3-level acyclic DAG
+(L1 `ht/client/` → L2 `ht/developer/` → L3 `services/`), with imports pointing
+strictly downward.
+
+## Deferred from `restructure-htas-3level-dag`
+
+These were captured (not done) by the structural-relocation change that moved
+`accessors.py` → `services/public/twin_io.py` and renamed the L1/L2 entry
+modules. They complete the move toward the clean single L2→L3 arrow.
+
+- **Fold `handlers.py` and `dispatch.py` into the L2 macro-verbs.** They are
+  currently *transitional L2-internal modules* sitting between the
+  `hydrological_twin_developer.py` facade and L3. Folding their compute /
+  routing bodies into the macro-verb methods removes the intermediate hop so L2
+  becomes a single layer that calls straight down into L3.
+- **De-orchestrate `operations_client.py`.** Today it owns full
+  `fetch → transform → render` chains (e.g. `run_mask_internal_values` ≈ 290
+  lines). Push that chaining down toward L2 macro-verbs so the L1 client stays
+  thin and the orchestration lives one level closer to the compute.
+- **Revisit moving the leaf data DTOs into an L3 `io_types.py`** — *done* for
+  the 5 read DTOs (`ValuesResponse`, `ObservationsResponse`, `CompartmentInfo`,
+  `LayerInfo`, `ObservationInfo`), which now live in
+  `services/public/io_types.py` and are re-exported by L2 `api_types.py`.
+  Revisit whether the remaining `*Response` / `*Request` DTOs that L3 never
+  needs should stay in L2, or whether more of them belong at the L3 leaf, once
+  `handlers.py` / `dispatch.py` are folded in.
 
 ## Spotted risks
 
@@ -39,7 +63,7 @@ smell is the unreached dispatch arm, which also still carries the 1-based-id /
 - **Fix:** drop the dead arm (or guard it explicitly as developer-only).
 
 ### 5. `_build_cells_gdf` mixes orchestration with pure assembly
-`ht/client/operations.py:975-980` (FIXME already in place). The function mixes
+`ht/client/operations_client.py:975-980` (FIXME already in place). The function mixes
 *selection* (`twin.mask` / `_build_outcropping_mesh_gdf` — orchestration, belongs
 in `run_mask_internal_values`) with pure GeoDataFrame *assembly* (gpd join +
 weight column — a services-layer op, no twin / no dispatch).
