+++ src/HydrologicalTwinAlphaSeries/ht/developer/api_types.py
@@ -6,6 +6,17 @@ from typing import Any, Dict, List, Optional, Union
 
 import numpy as np
 
+# Leaf data DTOs live at L3 (``services/public/io_types.py``) so the L3 io
+# service can import them with a strictly downward edge. They are re-exported
+# here so existing L2 import sites keep referencing them from ``api_types``.
+from ...services.public.io_types import (
+    CompartmentInfo,
+    LayerInfo,
+    ObservationInfo,
+    ObservationsResponse,
+    ValuesResponse,
+)
+
 # ---------------------------------------------------------------------------
 # Internal state model
 # ---------------------------------------------------------------------------
@@ -235,26 +246,6 @@ class ExportRequest:
 
 
 @dataclass
-class ValuesResponse:
-    """Per-cell time-series response shared by ``fetch`` and ``mask`` reads.
-
-    ``weights`` and ``clipped_geometries`` are populated on the weighted
-    polygon-mask path (``mask(kind="area_values", weighted=True)``) and on the
-    HYD ``resolution="reaches"`` path (where reaches are always boundary-clipped
-    and carry a length-fraction weight, even when ``weighted=False``). On every
-    other path both fields are ``None`` so the binary-mask response shape is
-    unchanged from the parent ``add-mask-macro`` capability.
-    """
-
-    data: np.ndarray
-    dates: np.ndarray
-    meta: Optional[Dict[str, Any]] = None
-    csv_path: Optional[str] = None
-    weights: Optional[np.ndarray] = None
-    clipped_geometries: Optional[List[Any]] = None
-
-
-@dataclass
 class TemporalOpResponse:
     data: np.ndarray
     date_labels: np.ndarray
@@ -265,13 +256,6 @@ class TemporalOpResponse:
 class SpatialAverageResponse:
     data: np.ndarray
     meta: Dict[str, Any]
-
-
-@dataclass
-class ObservationsResponse:
-    data: np.ndarray
-    dates: np.ndarray
-    meta: Optional[Dict[str, Any]] = None
 
 
 @dataclass
@@ -401,48 +385,6 @@ class AquiferBalanceResponse:
 # ---------------------------------------------------------------------------
 # Model and catalog responses
 # ---------------------------------------------------------------------------
-
-
-@dataclass
-class CompartmentInfo:
-    id_compartment: int
-    name: str
-    layers_gis_names: List[str]
-    n_layers: int
-    n_cells: int
-    cell_ids: np.ndarray
-    out_caw_path: str
-    regime: str
-    # 1-based global cell index (``Cell.id_abs``) in getCellIdVector order —
-    # the simulation-matrix row order. Distinct from ``cell_ids`` (per-layer).
-    id_abs: Optional[np.ndarray] = None
-
-
-@dataclass
-class LayerInfo:
-    id_layer: int
-    n_cells: int
-    cell_ids: np.ndarray
-    cell_areas: np.ndarray
-    cell_geometries: list
-    layer_gis_name: str
-    crs: Any = None
-    # 1-based global cell index (``Cell.id_abs``) for each cell in this layer.
-    id_abs: Optional[np.ndarray] = None
-
-
-@dataclass
-class ObservationInfo:
-    id_compartment: int
-    obs_type: str
-    n_points: int
-    layer_gis_name: str
-    point_names: List[str]
-    point_ids: list
-    cell_ids: List[int]
-    layer_ids: List[int]
-    geometries: list
-    mesh_ids: List[int]
 
 
 @dataclass
