+++ src/HydrologicalTwinAlphaSeries/services/public/__init__.py
@@ -8,9 +8,24 @@ from HydrologicalTwinAlphaSeries.services.public.geodata_assembly import (
     assemble_multi_layer_geodataframe,
     assemble_single_layer_geodataframe,
 )
+from HydrologicalTwinAlphaSeries.services.public.polygon_mask import (
+    aq_cells_boundary_faces,
+    aq_cells_on_polygon_boundary,
+    cells_in_polygon,
+    cells_in_polygon_weighted,
+    reaches_in_polygon_carachterisation,
+)
 from HydrologicalTwinAlphaSeries.services.public.renderer import Renderer
 from HydrologicalTwinAlphaSeries.services.public.spatial import Spatial
+from HydrologicalTwinAlphaSeries.services.public.io_types import (
+    CompartmentInfo,
+    LayerInfo,
+    ObservationInfo,
+    ObservationsResponse,
+    ValuesResponse,
+)
 from HydrologicalTwinAlphaSeries.services.public.temporal import CacheMissError, Temporal
+from HydrologicalTwinAlphaSeries.services.public import twin_io
 from HydrologicalTwinAlphaSeries.services.public.vec_operator import (
     Comparator,
     Extractor,
@@ -21,14 +36,25 @@ __all__ = [
     "Budget",
     "CacheMissError",
     "Comparator",
+    "CompartmentInfo",
     "DetectionError",
     "Extractor",
+    "LayerInfo",
+    "ObservationInfo",
+    "ObservationsResponse",
     "Operator",
     "Renderer",
     "Spatial",
     "Temporal",
+    "ValuesResponse",
+    "aq_cells_boundary_faces",
+    "aq_cells_on_polygon_boundary",
     "assemble_multi_layer_geodataframe",
     "assemble_single_layer_geodataframe",
+    "cells_in_polygon",
+    "cells_in_polygon_weighted",
     "detect_from_out_caw",
     "detect_project_neighbors",
+    "reaches_in_polygon_carachterisation",
+    "twin_io",
 ]
