+++ src/HydrologicalTwinAlphaSeries/tools/__init__.py
@@ -1,13 +1,10 @@
 from HydrologicalTwinAlphaSeries.tools.spatial_utils import (
     CRSMismatchError,
     SpatialIndex,
-    aq_cells_on_polygon_boundary,
-    cells_in_polygon,
     combine_geometries,
     get_nearest_cell,
     get_nearest_row,
     get_spatial_index,
-    reaches_in_polygon_carachterisation,
     read_hyd_corresp_file,
     reproject_to_match,
     verify_crs_match,
@@ -16,13 +13,10 @@ from HydrologicalTwinAlphaSeries.tools.spatial_utils import (
 __all__ = [
     "CRSMismatchError",
     "SpatialIndex",
-    "aq_cells_on_polygon_boundary",
-    "cells_in_polygon",
     "combine_geometries",
     "get_nearest_cell",
     "get_nearest_row",
     "get_spatial_index",
-    "reaches_in_polygon_carachterisation",
     "read_hyd_corresp_file",
     "reproject_to_match",
     "verify_crs_match",
